* implemented initial OpenXR input handling + updated event handling:    * event: use SDL_GetTicks64() instead of SDL_GetTicks() -> use 64-bit event times instead 32-bit everywhere    * added openxr_common.hpp that contains common OpenXR functionality (header includes, error handling, call wrappers)    * added openxr_input.cpp: this performs all OpenXR input handling (initial setup + bindings, updates, event emission)    * right now we can handle all hand/controller events, but some may need to be emulated in the future (note that this only has Valve Index controller bindings right now, others will follow)    * since OpenXR does not provide an action delta, all VR events have a zero delta for now (will possibly have to emulate this as well)    * openxr_context: get function pointers for XR_KHR_win32_convert_performance_counter_time and XR_KHR_convert_timespec_time    * openxr_context: added convert_time_to_ticks() helper function to convert an XrTime to SDL 64-bit ticks (milliseconds since start)    * openxr_context: added to_path()/to_path_or_throw() string -> XrPath conversion functions  * use SDL_GetTicks64() instead of SDL_GetTicks() everywhere  * bumped SDL2 requirement to 2.0.18+ (from 2.0.4)

+++ core/event.hpp
@@ -85,9 +85,9 @@ protected:
 	unordered_map<EVENT_TYPE, shared_ptr<event_object>> prev_events;
 	
 	//! timer that decides if there is a * mouse double click
-	uint32_t lm_double_click_timer;
-	uint32_t rm_double_click_timer;
-	uint32_t mm_double_click_timer;
+	uint64_t lm_double_click_timer;
+	uint64_t rm_double_click_timer;
+	uint64_t mm_double_click_timer;
 	
 	//! config setting for * mouse double click "timeframe"
 	uint32_t ldouble_click_time = 200;
